Changed integer value of data provider from 36 to 38

+++ src/main/java/qa/test/Data.java
@@ -8,7 +8,7 @@ public class Data
     @DataProvider( name = "dp" )
     public static Object[][] createData1() {
         return new Object[][] {
-                { "Iteration1", new Integer(36) },
+                { "Iteration1", new Integer(38) },
                 { "Iteration2", new Integer(37)},
         };
     }
